refactor: DTO를 사용하는 곳으로 패키지 변경
